Rename json test packages to reflect new import names
