+++ json/src/test/scala/busymachines/json_test/derive/JsonDefaultSemiAutoCodecDerivationTest.scala
@@ -1,4 +1,4 @@
-package busymachines.json_test.semiauto
+package busymachines.json_test.derive
 
 import busymachines.json_test._
 import org.scalatest.FlatSpec
