+++ json/src/test/scala/busymachines/json_test/derive/JsonDefaultSemiAutoDecoderDerivationTest.scala
@@ -1,12 +1,11 @@
-package busymachines.json_test.semiauto
+package busymachines.json_test.derive
 
 import busymachines.json_test._
 import org.scalatest.FlatSpec
 
 
 /**
-  *
-  * Here we test [[busymachines.json.Encoder]] derivation
+  * Here we test [[busymachines.json.Decoder]] derivation
   *
   * See the [[Melon]] hierarchy
   *
@@ -14,18 +13,16 @@ import org.scalatest.FlatSpec
   * @since 09 Aug 2017
   *
   */
-class JsonDefaultSemiAutoEncoderDerivationTest extends FlatSpec {
+class JsonDefaultSemiAutoDecoderDerivationTest extends FlatSpec {
 
   import busymachines.json.syntax._
-  import melonsDefaultSemiAutoEncoders._
+  import melonsDefaultSemiAutoDecoders._
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... be able to serialize anarchist melon (i.e. not part of any hierarchy)" in {
+  it should "... be able to deserialize anarchist melon (i.e. not part of any hierarchy)" in {
     val anarchistMelon = AnarchistMelon(noGods = true, noMasters = true, noSuperTypes = true)
-    val rawJson = anarchistMelon.asJson.spaces2
-
-    assertResult(
+    val rawJson =
       """
         |{
         |  "noGods" : true,
@@ -33,28 +30,28 @@ class JsonDefaultSemiAutoEncoderDerivationTest extends FlatSpec {
         |  "noSuperTypes" : true
         |}
       """.stripMargin.trim
-    )(rawJson)
+
+    val read = rawJson.unsafeDecodeAs[AnarchistMelon]
+    assertResult(anarchistMelon)(read)
   }
 
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... fail to compile when there is no defined encoder for a type down in the hierarchy" in {
+  it should "... fail to compile when there is no explicitly defined decoder for a type down in the hierarchy" in {
     assertDoesNotCompile(
       """
-        |val winterMelon: WinterMelon = WinterMelon(fuzzy = true, weight = 45)
-        |winterMelon.asJson
+        |val rawJson = "{}"
+        |rawJson.unsafeDecodeAs[WinterMelon]
       """.stripMargin
     )
   }
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... be able to serialize case class from hierarchy when it is referred to as its super-type" in {
+  it should "... be able to deserialize case class from hierarchy when it is referred to as its super-type" in {
     val winterMelon: Melon = WinterMelon(fuzzy = true, weight = 45)
-    val rawJson = winterMelon.asJson.spaces2
-
-    assertResult(
+    val rawJson =
       """
         |{
         |  "fuzzy" : true,
@@ -62,50 +59,51 @@ class JsonDefaultSemiAutoEncoderDerivationTest extends FlatSpec {
         |  "_type" : "WinterMelon"
         |}
       """.stripMargin.trim
-    )(rawJson)
+
+    val read = rawJson.unsafeDecodeAs[Melon]
+    assertResult(winterMelon)(read)
   }
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... be able to serialize case objects of the hierarchy" in {
+  it should "... be able to deserialize case objects of the hierarchy" in {
     val smallMelon: Melon = SmallMelon
-    val rawJson = smallMelon.asJson.spaces2
-    assertResult(
+    val rawJson =
       """
         |{
         |  "_type" : "SmallMelon"
         |}
       """.stripMargin.trim
-    )(rawJson)
+    val read = rawJson.unsafeDecodeAs[Melon]
+    assertResult(smallMelon)(read)
   }
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... serialize hierarchies of case objects as enums (i.e. plain strings)" in {
+  it should "... deserialize hierarchies of case objects as enums (i.e. plain strings)" in {
     val taste: List[Taste] = List(SweetTaste, SourTaste)
-
-    val rawJson = taste.asJson.spaces2
-    assertResult(
+    val rawJson =
       """
         |[
         |  "SweetTaste",
         |  "SourTaste"
         |]
       """.stripMargin.trim
-    )(rawJson)
+
+    val read = rawJson.unsafeDecodeAs[List[Taste]]
+    assertResult(read)(taste)
   }
 
   //-----------------------------------------------------------------------------------------------
 
-  it should "... serialize list of all case classes from the hierarchy" in {
+  it should "... deserialize list of all case classes from the hierarchy" in {
     val winterMelon: Melon = WinterMelon(fuzzy = true, weight = 45)
     val waterMelon: Melon = WaterMelon(seeds = true, weight = 90)
     val smallMelon: Melon = SmallMelon
     val squareMelon: Melon = SquareMelon(weight = 10, tastes = Seq(SourTaste, SweetTaste))
     val melons = List[Melon](winterMelon, waterMelon, smallMelon, squareMelon)
 
-    val rawJson = melons.asJson.spaces2
-    assertResult(
+    val rawJson =
       """
         |
         |[
@@ -132,9 +130,12 @@ class JsonDefaultSemiAutoEncoderDerivationTest extends FlatSpec {
         |  }
         |]
         |
-      """.stripMargin.trim
-    )(rawJson)
+        """.stripMargin.trim
+
+    val read: List[Melon] = rawJson.unsafeDecodeAs[List[Melon]]
+    assertResult(melons)(read)
   }
+
   //-----------------------------------------------------------------------------------------------
 
   //-----------------------------------------------------------------------------------------------
