+++ json/src/test/scala/busymachines/json_test/derive/semiAutoMelonJsonCodec.scala
@@ -1,4 +1,4 @@
-package busymachines.json_test.semiauto
+package busymachines.json_test.derive
 
 import busymachines.json_test._
